refactored cache with immediate eviction
