+++ flake-guard-app/src/server/server.ts
@@ -4,7 +4,7 @@ import bodyParser from 'body-parser';
 import cors from 'cors';
 require('dotenv').config();
 const app = express();
-import npmRouter from './routes/npm';
+import npmRouter from './routes/npmRouter';
 import tempDashRouter from './routes/tempDashRouter';
 
 const PORT = process.env.PORT || 3000;
